+++ src/apis/article/postArticles.ts
@@ -1,26 +1,22 @@
-"use server";
-
 import fetchInstance from "@/utils/fetchInstance";
 import { UserInput } from "@/types/auth";
 import { cookies } from "next/headers";
 
-const postSignIn = async (userInput: UserInput) => {
+const postArticles = async (userInput: UserInput) => {
   try {
-    const data = await fetchInstance("auth/signIn", {
+    const data = await fetchInstance("auth/signUp", {
       method: "POST",
       body: JSON.stringify(userInput),
     });
-
     cookies().set("cookie", data.accessToken);
-
     return data;
   } catch (error) {
     if (error instanceof Error) {
-      throw new Error(error.message || "login failed");
+      throw new Error(error.message || "sign up failed");
     } else {
-      throw new Error("login failed");
+      throw new Error("sign up failed");
     }
   }
 };
 
-export default postSignIn;
+export default postArticles;
